requirements.txt
PyPDF2==3.0.1
splitwise==2.5.0
numpy==1.24.3
python-telegram-bot==20.3
loguru=0.7.0